admin 相关查询接口

+++ overtime-common/common-authentication/src/main/java/org/overtime/common/auth/AuthInfo.java
@@ -14,7 +14,7 @@ public interface AuthInfo {
      *
      * @return id
      */
-    long getId();
+    Integer getId();
 
 
 }
